refactor(client): extract App.tsx hooks and inline components (#400)
Splits App.tsx (956 → 665 LOC) into focused, reusable units:
- `panel-layout.ts`: add PANEL_MIN/MAX/DEFAULT_WIDTH + loadPanelWidth()
- `components/EmptyState.tsx`: extract inline EmptyState component
- `components/ConnectionBanner.tsx`: extract inline ConnectionBanner component
- `components/PanelSlot.tsx`: ChatSlot + SideSlot with optional visible prop
- `hooks/useConnectionBanner.ts`: prolonged-disconnect banner state
- `hooks/useTandemModeBroadcast.ts`: tandem mode state, localStorage persist, Y.Map broadcast
- `hooks/useDragResize.ts`: panel drag-resize logic with atomic listener cleanup
