+++ src/client/panel-layout.ts
@@ -1,3 +1,5 @@
+import { PANEL_WIDTH_KEYS, type PanelSide } from "../shared/constants";
+
 /**
  * Client-local UI state describing the current side-panel arrangement.
  *
@@ -11,3 +13,26 @@
 export type PanelLayout =
   | { kind: "tabbed"; right: number }
   | { kind: "three-panel"; left: number; right: number };
+
+export const PANEL_MIN_WIDTH = 200;
+export const PANEL_MAX_WIDTH = 600;
+export const PANEL_DEFAULT_WIDTH = 300;
+
+export function loadPanelWidth(side: PanelSide): number {
+  const key = PANEL_WIDTH_KEYS[side];
+  try {
+    const saved = localStorage.getItem(key);
+    if (saved) {
+      const parsed = parseInt(saved, 10);
+      if (Number.isFinite(parsed)) {
+        return Math.max(PANEL_MIN_WIDTH, Math.min(PANEL_MAX_WIDTH, parsed));
+      }
+      // Non-finite saved value — fall through and warn so corrupt storage
+      // is diagnosable instead of silently reverting to the default.
+      console.warn(`[tandem] ignoring non-numeric panel width for ${key}: ${saved}`);
+    }
+  } catch (err) {
+    console.warn(`[tandem] localStorage unavailable reading ${key}:`, err);
+  }
+  return PANEL_DEFAULT_WIDTH;
+}
